requirements.txt
aioconsole==0.5.1
aiodns==3.0.0
aiofiles==0.6.0
aiohttp==3.8.1
aiohttp-cors==0.7.0
aiosignal==1.2.0
anyio==3.6.1
appdirs==1.4.4
arrow==1.2.3
astunparse==1.6.3
async-timeout==4.0.2
attrs==19.3.0
Automat==0.8.0
beautifulsoup4==4.11.1
bella==2.1.5
bernhard==0.2.6
better-ipc==1.1.4
betterstatcord.py==3.4.1
blinker==1.4
bottle==0.12.23
braceexpand==0.1.7
Brotli==1.0.9
bs4==0.0.1
button-paginator==1.0.1
cairocffi==1.4.0
CairoSVG==2.5.2
cassandra-driver==3.25.0
cchardet==2.1.7
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
chess==1.9.2
chevron==0.14.0
click==8.1.3
colorama==0.4.3
coloredlogs==15.0.1
configobj==5.0.6
constantly==15.1.0
CouchDB==1.2
cryptography==2.8
cssselect2==0.7.0
cycler==0.11.0
dataclass-factory==2.16
dataclasses-json==0.5.7
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.7.1
dnspython==2.2.1
docker==6.0.0
english-words==1.1.0
entrypoints==0.3
ffmpeg==1.4
fire==0.3.1
fonttools==4.37.1
frozenlist==1.3.1
future==0.18.2
fuzzywuzzy==0.18.0
geomet==0.2.1.post1
Glances==3.2.7
google-search-results==2.4.1
googletrans==3.0.0
graphitesender==0.11.2
greenlet==1.1.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.9.1
httpcore==0.9.1
httplib2==0.14.0
httpx==0.13.3
humanfriendly==10.0
humanize==4.3.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
ifaddr==0.2.0
imageio==2.21.2
imageio-ffmpeg==0.4.7
import-expression==1.1.4
importlib==1.0.4
importlib-metadata==4.12.0
lxml==4.9.1
marshmallow==3.16.0
marshmallow-enum==1.5.1
matplotlib==3.5.2
meson==0.53.2
motor==3.0.0
moviepy==1.0.3
msgpack==1.0.4
multidict==6.0.2
mypy-extensions==0.4.3
nest-asyncio==1.5.5
netifaces==0.10.4
numpy==1.22.3
oauthlib==3.2.0
package-name==0.1
packaging==21.3
pendulum==2.1.2
pika==1.3.0
Pillow==9.1.0
pipreqs==0.4.11
playwright==1.22.0
ply==3.11
potsdb==1.0.3
proglog==0.1.10
protobuf==4.21.7
psutil==5.9.1
py3nvml==0.2.7
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycares==4.2.2
pycparser==2.21
pycryptodomex==3.15.0
pydantic==1.10.2
pydub==0.24.1
pyee==8.1.0
PyNaCl==1.5.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyreadline3==3.4.1
pytz==2022.1
pytzdata==2020.1
PyYAML==5.3.1
pyzmq==24.0.1
reactivex==4.0.4
regex==2022.7.25
requests==2.27.1
requests-futures==1.0.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
scandir==1.10.0
service-identity==18.1.0
shazamio==0.1.1.1
simplejson==3.16.0
six==1.16.0
sos==4.3
soupsieve==2.3.2.post1
ssh-import-id==5.10
statcord.py==3.1.0
statsd==3.3.0
systemd-python==234
tabulate==0.8.10
termcolor==1.1.0
TikTokApi==5.1.1
timer==0.2.2
tinycss2==1.1.1
topggpy==1.4.0
tqdm==4.64.0
tweepy==4.10.0
Twisted==18.9.0
typing-extensions==4.2.0
typing-inspect==0.7.1
ua-parser==0.10.0
ubuntu-advantage-tools==27.10
ufw==0.36
ujson==4.3.0
unattended-upgrades==0.1
urbandict==0.6.1
urllib3==1.26.9
uvloop==0.17.0
validators==0.20.0
wadllib==1.3.3
Wand==0.6.10
webencodings==0.5.1
websocket-client==0.59.0
websockets==10.1
wifi==0.3.8
wordcloud==1.8.2.2
wrapt==1.14.1
xmltodict==0.13.0
yarg==0.1.9
yarl==1.7.2
youtube-dl==2021.12.17
yt-dlp==2022.9.1
ytdl==1.0.0rc5
zeroconf==0.39.1
zipp==3.8.0
zope.interface==5.4.0